+++ src/2d/time_integrator.cu
@@ -15,7 +15,7 @@ void checkCudaError(const char*);
  * 
  * I need to store u + alpha * k_i into some temporary variable called k*.
  */
-__global__ void rk4_tempstorage(float *c, float *kstar, float*k, float alpha, int n_p, int num_elem) {
+__global__ void rk4_tempstorage(double *c, double *kstar, double*k, double alpha, int n_p, int num_elem) {
     int idx = blockDim.x * blockIdx.x + threadIdx.x;
 
     if (idx < n_p * num_elem) {
@@ -28,7 +28,7 @@ __global__ void rk4_tempstorage(float *c, float *kstar, float*k, float alpha, in
  * computes the runge-kutta solution 
  * u_n+1 = u_n + k1/6 + k2/3 + k3/3 + k4/6
  */
-__global__ void rk4(float *c, float *k1, float *k2, float *k3, float *k4, int n_p, int num_elem) {
+__global__ void rk4(double *c, double *k1, double *k2, double *k3, double *k4, int n_p, int num_elem) {
     int idx = blockDim.x * blockIdx.x + threadIdx.x;
 
     if (idx < n_p * num_elem) {
@@ -42,12 +42,12 @@ __global__ void rk4(float *c, float *k1, float *k2, float *k3, float *k4, int n_
  * coefficients for each element
  * THREADS: num_elem
  */
-__global__ void eval_rhs_rk4(float *c, float *quad_rhs, float *left_riemann_rhs, float *right_riemann_rhs, 
+__global__ void eval_rhs_rk4(double *c, double *quad_rhs, double *left_riemann_rhs, double *right_riemann_rhs, 
                          int *elem_s1, int *elem_s2, int *elem_s3,
-                         int *left_elem, float *J, 
-                         float dt, int n_p, int num_sides, int num_elem) {
+                         int *left_elem, double *J, 
+                         double dt, int n_p, int num_sides, int num_elem) {
     int idx = blockDim.x * blockIdx.x + threadIdx.x;
-    float s1, s2, s3, register_J;
+    double s1, s2, s3, register_J;
     int i, s1_idx, s2_idx, s3_idx;
 
     if (idx < num_elem) {
@@ -86,28 +86,28 @@ __global__ void eval_rhs_rk4(float *c, float *quad_rhs, float *left_riemann_rhs,
     }
 }
 
-void time_integrate_rk4(float dt, int n_quad, int n_quad1d, int n_p, int n, 
+void time_integrate_rk4(double dt, int n_quad, int n_quad1d, int n_p, int n, 
                     int num_elem, int num_sides, int debug, int alpha, int timesteps) {
     int n_threads = 128;
     int i;
-    float t;
+    double t;
 
     int n_blocks_elem    = (num_elem  / n_threads) + ((num_elem  % n_threads) ? 1 : 0);
     int n_blocks_sides   = (num_sides / n_threads) + ((num_sides % n_threads) ? 1 : 0);
     int n_blocks_rk4     = ((n_p * num_elem) / n_threads) + (((n_p * num_elem) % n_threads) ? 1 : 0);
 
-    void (*eval_surface_ftn)(float*, float*, float*, 
-                         float*,
-                         float*, float*,
-                         float*, float*,
-                         float*, float*,
+    void (*eval_surface_ftn)(double*, double*, double*, 
+                         double*,
+                         double*, double*,
+                         double*, double*,
+                         double*, double*,
                          int*, int*,
                          int*, int*,
-                         float*, float*,
-                         int, int, int, int, float, int) = NULL;
-    void (*eval_volume_ftn)(float*, float*, 
-                        float*, float*, 
-                        float*, float*,
+                         double*, double*,
+                         int, int, int, int, double, int) = NULL;
+    void (*eval_volume_ftn)(double*, double*, 
+                        double*, double*, 
+                        double*, double*,
                         int, int, int) = NULL;
     switch (n) {
         case 0: eval_surface_ftn = eval_surface_wrapper0;
@@ -159,10 +159,10 @@ void time_integrate_rk4(float dt, int n_quad, int n_quad1d, int n_p, int n,
         if (debug) {
             printf("\n\n dt = %lf -\n", dt);
             printf("-------------------------\n");
-            float *left_rhs = (float *) malloc(num_sides * n_p * sizeof(float));
-            float *right_rhs = (float *) malloc(num_sides * n_p * sizeof(float));
-            cudaMemcpy(left_rhs, d_left_riemann_rhs, num_sides * n_p * sizeof(float), cudaMemcpyDeviceToHost);
-            cudaMemcpy(right_rhs, d_right_riemann_rhs, num_sides * n_p * sizeof(float), cudaMemcpyDeviceToHost);
+            double *left_rhs = (double *) malloc(num_sides * n_p * sizeof(double));
+            double *right_rhs = (double *) malloc(num_sides * n_p * sizeof(double));
+            cudaMemcpy(left_rhs, d_left_riemann_rhs, num_sides * n_p * sizeof(double), cudaMemcpyDeviceToHost);
+            cudaMemcpy(right_rhs, d_right_riemann_rhs, num_sides * n_p * sizeof(double), cudaMemcpyDeviceToHost);
             printf(" riemann\n");
             printf(" ~~~\n");
             for (int i = 0; i < num_sides * n_p; i++) {
@@ -184,8 +184,8 @@ void time_integrate_rk4(float dt, int n_quad, int n_quad1d, int n_p, int n,
         cudaThreadSynchronize();
 
         if (debug) {
-            float *quad_rhs = (float *) malloc(num_elem * n_p * sizeof(float));
-            cudaMemcpy(quad_rhs, d_quad_rhs, num_elem * n_p * sizeof(float), cudaMemcpyDeviceToHost);
+            double *quad_rhs = (double *) malloc(num_elem * n_p * sizeof(double));
+            cudaMemcpy(quad_rhs, d_quad_rhs, num_elem * n_p * sizeof(double), cudaMemcpyDeviceToHost);
             printf(" quad_rhs\n");
             printf(" ~~~\n");
             for (int i = 0; i < num_elem * n_p; i++) {
@@ -203,8 +203,8 @@ void time_integrate_rk4(float dt, int n_quad, int n_quad1d, int n_p, int n,
         cudaThreadSynchronize();
 
         if (debug) {
-            float *rhs = (float *) malloc(num_elem * n_p * sizeof(float));
-            cudaMemcpy(rhs, d_k1, num_elem * n_p * sizeof(float), cudaMemcpyDeviceToHost);
+            double *rhs = (double *) malloc(num_elem * n_p * sizeof(double));
+            cudaMemcpy(rhs, d_k1, num_elem * n_p * sizeof(double), cudaMemcpyDeviceToHost);
             printf(" eval_rhs\n");
             printf(" ~~~\n");
             for (int i = 0; i < num_elem * n_p; i++) {
@@ -314,12 +314,12 @@ void time_integrate_rk4(float dt, int n_quad, int n_quad1d, int n_p, int n,
  * FORWARD EULER
  ***********************/
 
-__global__ void eval_rhs_fe(float *c, float *quad_rhs, float *left_riemann_rhs, float *right_riemann_rhs, 
+__global__ void eval_rhs_fe(double *c, double *quad_rhs, double *left_riemann_rhs, double *right_riemann_rhs, 
                          int *elem_s1, int *elem_s2, int *elem_s3,
-                         int *left_elem, float *J, 
-                         float dt, int n_p, int num_sides, int num_elem) {
+                         int *left_elem, double *J, 
+                         double dt, int n_p, int num_sides, int num_elem) {
     int idx = blockDim.x * blockIdx.x + threadIdx.x;
-    float s1, s2, s3, register_J;
+    double s1, s2, s3, register_J;
     int i, s1_idx, s2_idx, s3_idx;
 
     if (idx < num_elem) {
@@ -359,27 +359,27 @@ __global__ void eval_rhs_fe(float *c, float *quad_rhs, float *left_riemann_rhs, 
 }
 
 // forward eulers
-void time_integrate_fe(float dt, int n_quad, int n_quad1d, int n_p, int n, 
+void time_integrate_fe(double dt, int n_quad, int n_quad1d, int n_p, int n, 
               int num_elem, int num_sides, int alpha, int timesteps) {
     int n_threads = 128;
     int i;
-    float t;
+    double t;
 
     int n_blocks_elem    = (num_elem  / n_threads) + ((num_elem  % n_threads) ? 1 : 0);
     int n_blocks_sides   = (num_sides / n_threads) + ((num_sides % n_threads) ? 1 : 0);
 
-    void (*eval_surface_ftn)(float*, float*, float*, 
-                         float*,
-                         float*, float*,
-                         float*, float*,
-                         float*, float*,
+    void (*eval_surface_ftn)(double*, double*, double*, 
+                         double*,
+                         double*, double*,
+                         double*, double*,
+                         double*, double*,
                          int*, int*,
                          int*, int*,
-                         float*, float*,
-                         int, int, int, int, float, int) = NULL;
-    void (*eval_volume_ftn)(float*, float*, 
-                        float*, float*, 
-                        float*, float*,
+                         double*, double*,
+                         int, int, int, int, double, int) = NULL;
+    void (*eval_volume_ftn)(double*, double*, 
+                        double*, double*, 
+                        double*, double*,
                         int, int, int) = NULL;
     switch (n) {
         case 0: eval_surface_ftn = eval_surface_wrapper0;
